requirements.txt
Flask==3.1.2
flask_clacks==1.0.1
flask_socketio==5.6.0
geocoder==1.38.1
psycopg2==2.9.11
python-dotenv==1.2.1
Requests==2.32.5
Werkzeug==3.1.4
